refactor: isolate Calendar stories
